Modified HelperFunctions::string_replace to html decode strings; modified SendMessages to use it.
I've added a default boolean parameter to string_replace to indicate whether the strings in the
tuple should be html decoded or not. By default they are not, which maintains existing behaviour.

All information that is stored in the database should be html encoded. SendMessages was retrieving
data from the database that was html encoded and contained html entities within it.
This is fine when sending HTML emails as the html encoded entities are displayed as intended.
However, when sending plain text emails the entities are, of course, displayed in their encoded form.
When populating plain text emails, TRUE can be passed to string_replace to decode html entities back
into characters.